+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Wrap the chart constructor, so you do not need to call `clj->js`/`#js`.
+
 # 0.1.0 (2016-08-04)
 
 Initial release.